move value types constant to formatter util file
