--- packages/utils/src/constant.js
@@ -1,16 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-
-export const VALUE_TYPES = {
-  BOOLEAN: 'boolean',
-  CURRENCY: 'currency',
-  FRACTION: 'fraction',
-  PERCENTAGE: 'percentage',
-  FRACTION_AND_PERCENTAGE: 'fractionAndPercentage',
-  NUMBER_AND_PERCENTAGE: 'numberAndPercentage',
-  TEXT: 'text',
-  NUMBER: 'number',
-  ONE_DECIMAL_PLACE: 'oneDecimalPlace',
-};
