+++ packages/utils/src/formatDataValueByType.js
@@ -4,7 +4,18 @@
  */
 
 import numeral from 'numeral';
-import { VALUE_TYPES } from './constant';
+
+export const VALUE_TYPES = {
+  BOOLEAN: 'boolean',
+  CURRENCY: 'currency',
+  FRACTION: 'fraction',
+  PERCENTAGE: 'percentage',
+  FRACTION_AND_PERCENTAGE: 'fractionAndPercentage',
+  NUMBER_AND_PERCENTAGE: 'numberAndPercentage',
+  TEXT: 'text',
+  NUMBER: 'number',
+  ONE_DECIMAL_PLACE: 'oneDecimalPlace',
+};
 
 /**
  *
