+++ packages/utils/src/index.js
@@ -26,5 +26,4 @@ export { WorkBookParser } from './WorkBookParser';
 export { checkValueSatisfiesCondition } from './checkValueSatisfiesCondition';
 export { addExportedDateAndOriginAtTheSheetBottom } from './addExportDateAndOriginInExcelExportData';
 export { getBrowserTimeZone } from './getBrowserTimeZone';
-export { formatDataValueByType } from './formatDataValueByType';
-export { VALUE_TYPES } from './constant';
+export { VALUE_TYPES, formatDataValueByType } from './formatDataValueByType';
